Support forcing layernorm to run in FP32 during ONNX export

+++ onnx_exporter/efficientvit_sam_onnx.py
@@ -66,7 +66,7 @@ class EfficientSamOnnxModel(nn.Module):
         input_image_size = input_image_size.to(torch.float32)
         scale = longest_side / torch.max(input_image_size)
         transformed_size = scale * input_image_size
-        transformed_size = torch.floor(transformed_size + 0.5).to(torch.int64)
+        transformed_size = torch.floor(transformed_size + 0.5).to(torch.int32)
         return transformed_size
 
     def _embed_points(self, point_coords: torch.Tensor, point_labels: torch.Tensor) -> torch.Tensor:
@@ -104,7 +104,6 @@ class EfficientSamOnnxModel(nn.Module):
 
         prepadded_size = self.resize_longest_image_size(orig_im_size, self.img_size)
         masks = masks[..., : int(prepadded_size[0]), : int(prepadded_size[1])]
-        orig_im_size = orig_im_size.to(torch.int64)
         h, w = orig_im_size[0], orig_im_size[1]
         masks = F.interpolate(masks, size=(h, w), mode="bilinear", align_corners=False)
         return masks
